refactor: consolidate data-model types into LDAIConfigTypes (AIC-2662)
Group the shared AI Config shapes (Mode, Message, Model, Provider, Tool,
JudgeConfiguration) as nested types under a single non-instantiable
LDAIConfigTypes container instead of separate top-level classes. This
reduces the file count and frees up generic names like Message and Model
within the datamodel package, matching the organization used by the
.NET SDK (LdAiConfigTypes). Per Matt's review on #171.

Internal references (AIConfigParser, AIConfigFlagValue) and tests are
updated to the nested names; behavior is unchanged.
